add auth system to each request

+++ backend/AirlineSurveys/surveysApp/views.py
@@ -75,6 +75,8 @@ def choose_choice(request) :
 
 
 def get_all_passengers(request, manager_id):
+    if not jwt_auth(request):
+        return HttpResponse("UnAuthorized!", status=401)
     if request.method == "GET":
         passengers = service.get_all_passengers(manager_id)
         return HttpResponse(json.dumps(passengers, indent=2), status=200)
@@ -82,6 +84,8 @@ def get_all_passengers(request, manager_id):
 
 
 def get_manager_surveys(request, manager_id):
+    if not jwt_auth(request):
+        return HttpResponse("UnAuthorized!", status=401)
     if request.method == "GET":
         surveys = service.get_surveys(manager_id)
         return HttpResponse(
@@ -114,6 +118,7 @@ def login(request):
     return HttpResponse("Method not allowed", status=405)
 
 
+
 @csrf_exempt
 def get_answers_by_number(request , sid , qnum) : 
     if request.method == "GET":
@@ -125,6 +130,8 @@ def get_answers_by_number(request , sid , qnum) :
 
 @csrf_exempt
 def survey(request, sid):
+    if not jwt_auth(request):
+        return HttpResponse("UnAuthorized!", status=401)
     try:
         if request.method == "GET":
             survey_info = service.get_survey_info(sid)
@@ -176,6 +183,9 @@ def question(request, surveyid, qnumber):
 
 @csrf_exempt
 def question_delete(request, sid, qid):
+    if not jwt_auth(request):
+        return HttpResponse("UnAuthorized!", status=401)
+
     if request.method == "DELETE":
         response = service.delete_question(sid, qid)
         print(f'question response: {response}')
@@ -191,6 +201,9 @@ def question_delete(request, sid, qid):
 
 @csrf_exempt
 def question_edit(request, sid, qid):
+    if not jwt_auth(request):
+        return HttpResponse("UnAuthorized!", status=401)
+
     if request.method == "PUT":
         data = json.loads(request.body)
         response = service.update_question(sid, qid, data)
@@ -206,9 +219,11 @@ def question_edit(request, sid, qid):
 
 def jwt_auth(request):
     token = request.META.get('HTTP_AUTHORIZATION')
-    token = token.split()[-1]
+    if token is None:
+        return False
 
     try:
+        token = token.split()[-1]
         jwt.decode(token, 'secret', algorithms=['HS256'])
         return True
     except:
